refactor: 重构Avatar.cs，将角色特化逻辑与战斗识别提取到单独文件
- AvatarSpecialActions.cs：特化动作注册表 + ExecuteSpecializedAction 分派

- AvatarRecognition.cs：FindBloodBars、FindDamageNumber 等通用识别方法

- 使用统一的 ActionArgs record 替代多个参数类型，用命名参数区分 UseSkill/Charge

- SpecializedRules 字典管理特化规则，支持参数条件预检
